+++ src/main/java/com/owczarczak/footballers/footballer/FootballerDto.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @Builder
 public class FootballerDto {
-    private int id;
+    private Long id;
     private String pesel;
     private String name;
-    private int height;
+    private Integer height;
 }
